+++ Models/PostModel/PostModel.php
@@ -173,6 +173,26 @@ abstract class PostModel
 	}
 
 	/**
+	 * Execute PostModel::where with meta_relation AND
+	 *
+	 * @see PostModel::where();
+	 *
+	 * @example ExamplePost::where('active', 1)->andWhere('spotlight', 'front')->get();
+	 *
+	 * @param string $meta_key
+	 * @param string $meta_value
+	 * @param string $meta_compare
+	 *
+	 * @return $this
+	 */
+	protected function andWhere($meta_key, $meta_value, $meta_compare = '=')
+	{
+		$this->where($meta_key, $meta_value, $meta_compare, 'AND');
+
+		return $this;
+	}
+
+	/**
 	 * Return Posts where they are by a certain author
 	 *
 	 * @see https://developer.wordpress.org/reference/classes/wp_query/#author-parameters
@@ -792,6 +812,7 @@ abstract class PostModel
 	 * @return $this
 	 */
 	private function appendDate()
+	{
 	{
 		foreach ($this->query->posts as $post) {
 			if (is_object($post)) {